requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dateutil==2.8.2
Werkzeug==2.3.7
redis==4.5.5
